refactor(typedMap): Extract and rename typedMap function into NativeSerializer module
Rather than exposing typedMap on the Bugsnag module in order to be able to test it, it made sense
for it to be in its own file. This refactor pulls it out, gives it a sensible name and then tests it
in isolation.
