+++ src/Bugsnag.js
@@ -1,6 +1,7 @@
 /* global ErrorUtils, __DEV__ */
 
 import { NativeModules } from 'react-native'
+import serializeForNativeLayer from './NativeSerializer'
 
 const NativeClient = NativeModules.BugsnagReactNative
 
@@ -159,7 +160,7 @@ export class Client {
     NativeClient.leaveBreadcrumb({
       name,
       type,
-      metadata: typedMap(breadcrumbMetaData)
+      metadata: serializeForNativeLayer(breadcrumbMetaData)
     })
   }
 
@@ -353,7 +354,7 @@ export class Report {
       errorClass: this.errorClass,
       errorMessage: this.errorMessage,
       groupingHash: this.groupingHash,
-      metadata: typedMap(this.metadata),
+      metadata: serializeForNativeLayer(this.metadata),
       severity: this.severity,
       stacktrace: this.stacktrace,
       user: this.user,
@@ -363,42 +364,3 @@ export class Report {
     }
   }
 }
-
-const allowedMapObjectTypes = ['string', 'number', 'boolean']
-
-/**
- * Convert an object into a structure with types suitable for serializing
- * across to native code.
- */
-const typedMap = function (map, maxDepth = 10, depth = 0, seen = new Set()) {
-  seen.add(map)
-  const output = {}
-  for (const key in map) {
-    if (!{}.hasOwnProperty.call(map, key)) continue
-
-    const value = map[key]
-
-    // Checks for `null`, NaN, and `undefined`.
-    if ([ undefined, null ].includes(value) || (typeof value === 'number' && isNaN(value))) {
-      output[key] = {type: 'string', value: String(value)}
-    } else if (typeof value === 'object') {
-      if (seen.has(value)) {
-        output[key] = {type: 'string', value: '[circular]'}
-      } else if (depth === maxDepth) {
-        output[key] = {type: 'string', value: '[max depth exceeded]'}
-      } else {
-        output[key] = {type: 'map', value: typedMap(value, maxDepth, depth + 1, seen)}
-      }
-    } else {
-      const type = typeof value
-      if (allowedMapObjectTypes.includes(type)) {
-        output[key] = {type: type, value: value}
-      } else {
-        console.warn(`Could not serialize breadcrumb data for '${key}': Invalid type '${type}'`)
-      }
-    }
-  }
-  return output
-}
-
-export { typedMap }
