+++ src/__tests__/Bugsnag.test.js
@@ -358,137 +358,3 @@ test('{enable|disable}ConsoleBreadCrumbs(): gracefully handles serialization edg
   // switched off for the next test ¯\_(ツ)_/¯
   c.disableConsoleBreadCrumbs()
 })
-
-test('typedMap converts arrays, objects, strings and numbers', () => {
-  const { typedMap } = require('../Bugsnag')
-  expect(typedMap({
-    arr: [ 0, 1, 2 ],
-    obj: {
-      a: 'a',
-      b: 'b',
-      c: 'c'
-    },
-    x: 10,
-    y: 'str'
-  })).toEqual({
-    arr: {
-      type: 'map',
-      value: {
-        '0': { type: 'number', value: 0 },
-        '1': { type: 'number', value: 1 },
-        '2': { type: 'number', value: 2 }
-      }
-    },
-    obj: {
-      type: 'map',
-      value: {
-        a: { type: 'string', value: 'a' },
-        b: { type: 'string', value: 'b' },
-        c: { type: 'string', value: 'c' }
-      }
-    },
-    x: { type: 'number', value: 10 },
-    y: { type: 'string', value: 'str' }
-  })
-})
-
-test('typedMap handles circular structures', () => {
-  const { typedMap } = require('../Bugsnag')
-  const objA = {}
-  objA.ref = objA
-
-  // this tests the top level object being referred to as a node
-  expect(typedMap(objA)).toEqual({ ref: { type: 'string', value: '[circular]' } })
-
-  const objB = {}
-  objB.leafA = {}
-  objB.leafB = {}
-  objB.leafA.ref = objB.leafB
-  objB.leafB.ref = objB.leafA
-
-  // this tests subtrees referring to eachother
-  expect(typedMap(objB)).toEqual({
-    leafA: {
-      type: 'map',
-      value: {
-        ref: {
-          type: 'map',
-          value: {
-            ref: { type: 'string', value: '[circular]' }
-          }
-        }
-      }
-    },
-    leafB: { type: 'string', value: '[circular]' }
-  })
-})
-
-test('typedMap handles deep nesting', () => {
-  const { typedMap } = require('../Bugsnag')
-  const nesty = {
-    zero: {
-      one: {
-        two: {
-          three: {
-            four: {
-              five: { six: { seven: { eight: { nine: { ten: { eleven: 'TOO_DEEP!' } } } } } }
-            }
-          }
-        }
-      }
-    }
-  }
-  expect(typedMap(nesty)).toEqual({
-    zero: {
-      type: 'map',
-      value: {
-        one: {
-          type: 'map',
-          value: {
-            two: {
-              type: 'map',
-              value: {
-                three: {
-                  type: 'map',
-                  value: {
-                    four: {
-                      type: 'map',
-                      value: {
-                        five: {
-                          type: 'map',
-                          value: {
-                            six: {
-                              type: 'map',
-                              value: {
-                                seven: {
-                                  type: 'map',
-                                  value: {
-                                    eight: {
-                                      type: 'map',
-                                      value: {
-                                        nine: {
-                                          type: 'map',
-                                          value: {
-                                            ten: { type: 'string', value: '[max depth exceeded]' }
-                                          }
-                                        }
-                                      }
-                                    }
-                                  }
-                                }
-                              }
-                            }
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  })
-})
